Move helper methods from EliminateCrossJoins optimizer to rule
